+++ log/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2018-12-08
+
+- remove logger relation ship from logger struct, use logger registry instead
+
 ## 2018-11-25
 
 - reduce `Handler` interface method from 6 to 1, just a `HandleLog(level Level, now time.Time, msg string, source string, context Fields, fields Fields)`